Add exparimental logger that's called in robot periodic

+++ src/main/java/frc/robot/lib/EncoderLib.java
@@ -61,7 +61,7 @@ public class EncoderLib extends Subsystem {
      * @return distance
      * 
      */
-    public static double rawToDistance(int rawPosition, double unitsPerRotation, double effectiveDiam) {
+    public static double rawToDistance(double rawPosition, double unitsPerRotation, double effectiveDiam) {
         double rotations = rawPosition / unitsPerRotation;
         double circumference = Math.PI * effectiveDiam;
         double distance = rotations * circumference;
